+++ src/rust/src/x509/certificate.rs
@@ -11,7 +11,7 @@ use cryptography_x509::extensions::{
     Admission, Admissions, AuthorityKeyIdentifier, BasicConstraints, DisplayText,
     DistributionPoint, DistributionPointName, DuplicateExtensionsError, ExtendedKeyUsage,
     Extension, IssuerAlternativeName, KeyUsage, MSCertificateTemplate, NameConstraints,
-    NamingAuthority, PolicyConstraints, PolicyInformation, PolicyQualifierInfo,
+    NamingAuthority, PolicyConstraints, PolicyInformation, PolicyMappings, PolicyQualifierInfo,
     PrivateKeyUsagePeriod, ProfessionInfo, Qualifier, RawExtensions, SequenceOfAccessDescriptions,
     SequenceOfSubtrees, SubjectAlternativeName, UserNotice,
 };
@@ -888,6 +888,17 @@ pub fn parse_cert_ext<'p>(
                 pc.require_explicit_policy,
                 pc.inhibit_policy_mapping,
             ))?))
+        }
+        oid::POLICY_MAPPINGS_OID => {
+            let mappings = ext.value::<PolicyMappings<'_, Asn1Read>>()?;
+            let py_mappings = pyo3::types::PyList::empty(py);
+            for mapping in mappings {
+                py_mappings.append((
+                    oid_to_py_oid(py, &mapping.issuer_domain_policy)?,
+                    oid_to_py_oid(py, &mapping.subject_domain_policy)?,
+                ))?;
+            }
+            Ok(Some(types::POLICY_MAPPINGS.get(py)?.call1((py_mappings,))?))
         }
         oid::OCSP_NO_CHECK_OID => {
             ext.value::<()>()?;